[FEATURE][REFACTOR] Implement post page and restructure module
