+++ src/app/[locale]/(protected)/(.)posts/[id]/loading.tsx
@@ -1,4 +1,4 @@
-import { PostModalSkeleton } from '@/modules/posts/post/components/post';
+import { PostModalSkeleton } from '@/modules/posts/post/components/post/modal';
 
 const PostModalLoadingPage = () => <PostModalSkeleton />;
