+++ src/modules/posts/post/components/post/index.ts
@@ -1,4 +1,1 @@
 export { Post } from './Post';
-export { PostModalSkeleton } from './modal/skeleton/PostModalSkeleton';
-export { PostModalWrapper } from './modal/wrapper/PostModalWrapper';
-export { PostSkeleton } from './skeleton/PostSkeleton';
