+++ src/modules/posts/post/components/post/layout/common/index.ts
@@ -1,0 +1,2 @@
+export { PostOptions } from './options/PostOptions';
+export { PostSkeleton } from './skeleton/PostSkeleton';
