+++ src/modules/posts/post/components/post/layout/common/skeleton/index.ts
