+++ src/modules/posts/post/components/post/layout/footer/comments/index.ts
