+++ src/modules/posts/post/components/post/layout/footer/info/PostFooterInfo.tsx
