+++ src/modules/posts/post/components/post/layout/index.ts
@@ -1,0 +1,3 @@
+export { PostBody } from './body/PostBody';
+export { PostFooter } from './footer/PostFooter';
+export { PostHeader } from './header/PostHeader';
