+++ src/modules/posts/post/components/post/modal/index.ts
@@ -1,0 +1,2 @@
+export { PostModalSkeleton } from './skeleton/PostModalSkeleton';
+export { PostModalWrapper } from './wrapper/PostModalWrapper';
