+++ src/modules/posts/post/components/post/page/index.ts
@@ -1,0 +1,1 @@
+export { PostPageWrapper } from './wrapper/PostPageWrapper';
