+++ src/modules/posts/post/components/post/shared/index.ts
@@ -1,0 +1,1 @@
+export { PostContent } from './content/PostContent';
