+++ src/modules/posts/post/components/posts/PostsSkeleton.tsx
@@ -1,4 +1,4 @@
-import { PostSkeleton } from './post';
+import { PostSkeleton } from '../post/layout/common';
 
 export const PostsSkeleton = () => {
   return [...Array(5)].map((_, index) => <PostSkeleton key={index} />);
