+++ src/modules/posts/post/components/posts/index.ts
